Bring the Worker class back
