+++ crawler/vms_crawler.py
@@ -1,5 +1,5 @@
 import plugins_manager
-from base_crawler import Crawler, BaseFrame
+from base_crawler import BaseCrawler, BaseFrame
 from virtual_machine import get_virtual_machines
 
 
@@ -11,21 +11,17 @@ class VirtualMachineFrame(BaseFrame):
         self.metadata['system_type'] = 'vm'
 
 
-class VirtualMachinesCrawler(Crawler):
+class VirtualMachinesCrawler(BaseCrawler):
 
     def __init__(self,
-                 emitters=None,
-                 frequency=-1,
                  features=['os', 'cpu'],
                  user_list=[],
                  host_namespace='',
                  plugin_places=['plugins'],
                  options={}):
 
-        Crawler.__init__(
+        BaseCrawler.__init__(
             self,
-            emitters=emitters,
-            frequency=frequency,
             features=features,
             plugin_places=plugin_places,
             options=options)
